Introduce a CursorClient trait for the cursor RPC surface
Move the cursor RPC surface out of tds_client.rs into a new connection/cursor_ops.rs and expose it as an async CursorClient trait implemented for TdsClient (mirroring ResultSetClient), so cursor behavior is a distinct, swappable abstraction rather than inherent methods on the client.

The 11 sp_cursor* methods form the trait contract; their cursor-only helpers (extract_return_value_int, consume_cursor_open_response, captured_cursor_status, drain_cursor_response) remain private inherent methods. The shared helpers the impl reuses (check_and_reconnect, deduct_timeout, timeout_to_duration, drain_stream), the ReturnStatus enum, and the fields the cursor methods touch are pub(crate). Method bodies are unchanged, so there is no behavioral change.

Implements AB#45976.
